Added all necessary functionality to Calendar, Week and Event (primarily stuff related to Ideal Weeks, both for display and for recurrent events). Added all helper methods needed for easy implementation of all types of event (particularly recurrent events, since those need to track themselves in special lists. Riley knows how this needs to be implemented). This should be everything we need in entities.
I have mostly ignored EventFactory. I just wanted to get the necessary functionality out, since we need to move beyond entities. There is a possibliity that a more clever design pattern is possible, particularly one that uses an abstract factory and corresponding interfaces to differentiate between Commutes and Normal events for every type of event. This would make some of the methods a bit less awkward. But we can retroactively update this later, as of now we really just need this functionality.

Documentation is there, but might be lacking.

Safety checks also need to be added. Several of these methods should throw different exceptions, and as of right now they simply try to do whatever and then everything explodes. Will add this later as well.