death has come for us all.
implementation had to radically change to reflect clean architecture. the lack of permanently accessible instances of entities means that most of our pre-existing code does not work. use case interactors need to deal with the CSVs directly, and this code reflects that.

if we eventually decide that we want to revive the calendar class, then calendar 2.0 is the most updated version of that conceptualization of the code. good luck everyone